requirements.txt
flask
flask-cors
gunicorn
python-dotenv
requests
tenacity
openai
moviepy<2.0.0
mutagen
pypdf
Pillow
jsonschema
edge-tts
pydub
manim
pydantic
rembg
pyttsx3
numpy
